+++ test/Main.hs
@@ -10,6 +10,7 @@ import Pudu.Frontend.Lexer.NumberSymbolSpec (numberSymbolProperties)
 import Pudu.Frontend.Lexer.QuotedSpec (quotedProperties)
 import Pudu.Frontend.Lexer.ScannerSpec (scannerProperties)
 import Pudu.Frontend.ParserStateNameSpec (parserStateNameProperties)
+import Pudu.Frontend.ParserExpressionSpec (parserExpressionProperties)
 import Pudu.Frontend.ParserTypeSpec (parserTypeProperties)
 import Pudu.Frontend.SyntaxSpec (syntaxProperties)
 import Pudu.Frontend.TokenSpec (tokenProperties)
@@ -45,7 +46,8 @@ main = do
   syntaxOutcomes <- traverse (uncurry check) syntaxProperties
   parserStateNameOutcomes <- traverse (uncurry check) parserStateNameProperties
   parserTypeOutcomes <- traverse (uncurry check) parserTypeProperties
-  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserTypeOutcomes)) exitFailure
+  parserExpressionOutcomes <- traverse (uncurry check) parserExpressionProperties
+  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserTypeOutcomes <> parserExpressionOutcomes)) exitFailure
 check :: String -> IO Property -> IO Bool
 check label loadProperty = do
   putStrLn ("[test] " <> label)